non static context in cracker-tool

+++ Cracker.h
@@ -13,7 +13,7 @@ public:
     bool cracker( byte y[], Rainbow& rainbow, Context& context,
                   string & clear );
     bool verifyAlert( byte hashed[], Context context, vector<Chain> X, int t, int m, string& clear);
-    uint getProbability(uint64 N, uint M, uint T);
+    double getProbability(uint64 N, uint M, uint T);
 };
